+++ apps/api/src/wordstranslation/wordstranslation.entity.ts
@@ -3,13 +3,13 @@ import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
 
 @Entity({ name: "words_translation" })
 export class WordTranslationEntity implements WordTranslation {
-	@PrimaryGeneratedColumn({ type: "bigint" })
+	@PrimaryGeneratedColumn({ type: "int" })
 	id: number;
 
 	@Column({ type: "timestamptz" })
 	created_at: string;
 
-	@Column({ type: "bigint" })
+	@Column({ type: "int" })
 	word: number;
 
 	@Column()
